fix(a11y): focus on congratulations content when form successful
This is a really exciting commit to me! Despite being a few lines of code,
it adds something I think is really valuable.

Our problem is this: when the form is submitted, new content becomes visible
if your submission is successful. If you are using an assistive device, like
VoiceOver: this new content is only apparent to you if you continue to scroll
down the page.

An accessible experience would be to immediately alert this new content to the user.
In order to do this, within the form JavaScript (`onSuccess`), I have set the
new content to be focused on. If using something like voice-over (try it!),
this new content will begin to be immediately announced!

In order for this content block to the focusable, I have added a tab index of '-1',
this denotes an element focusable by a script but not the user.

I have also removed the outline that comes with focusing on this element, as it
is not a necessary visual enhancement.